Another revamp of structure | create a package for global types
